requirements.txt
GitPython==2.1.10
nltk==3.3